pyproject.toml
[tool.poetry]
name = "furuta"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "furuta"}]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
gymnasium = "^0.29.1"
pyserial = "^3.5"
simple-pid = "^2.0.0"
mcap-protobuf-support = "^0.4.1"
scipy = "^1.11.4"
pygame = "^2.5.2"
pin = "^2.7.0"
crocoddyl = "^2.0.2"
panda3d-viewer = "^0.4.1"
cmeel-mim-solvers = "^0.0.4"
onshape-to-robot = "^0.3.26"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"
matplotlib = "3.7.3"
pre-commit = "^3.5.0"
stable-baselines3 = "^2.2.1"
wandb = "^0.16.2"
opencv-python = "^4.9.0.80"
hydra-core = "^1.3.2"
tensorboard = "^2.15.1"
moviepy = "^1.0.3"
sbx-rl = "^0.10.0"
jax-metal = {version = "^0.0.5", platform = "darwin"}
sb3-contrib = {git = "https://github.com/Stable-Baselines-Team/stable-baselines3-contrib"}
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
